app/requirements.txt
torch>=2.0
torchvision
Pillow
fastapi
uvicorn
openai
langchain
python-dotenv
sentence-transformers
faiss-cpu
crewai
numpy
scikit-learn
python-multipart
langchain-community
pypdf
langchain-huggingface
opencv-python
matplotlib
langchain-openai
streamlit
langchain_google_genai
google-genai